requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.39
Werkzeug==3.1.3
pandas==2.2.3
openpyxl==3.1.5
FPDF==1.7.2
python-dotenv==1.0.1
Flask-WTF==1.2.2
flask-login==0.6.3
jinja2==3.1.6
weasyprint==65.0
wtforms==3.2.1
trafilatura==2.0.0
oauthlib==3.2.2
pillow==11.1.0
trafilatura
flask-login==0.6.3
pdfkit==1.0.0
email_validator==2.2.0
gunicorn==23.0.0
psycopg2-binary==2.9.10
email-validator
flask
flask-login
flask-sqlalchemy
flask-wtf
gunicorn
jinja2
openpyxl
pandas
psycopg2-binary
sqlalchemy
trafilatura
weasyprint
werkzeug
wtforms
